hyblock_dashboard/requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1

# Monitoring and utilities
psutil==5.9.8
watchdog==3.0.0

# Data processing
pytz==2023.3
tzlocal==5.2
jsonschema==4.20.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Development tools
black==23.12.1
flake8==7.0.0
pytest==7.4.4 